backend/services/location-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets<13
pydantic==2.6.0
pydantic-settings==2.1.0
redis==5.0.1                           # ElastiCache — GPS caching + pub/sub
pytest-asyncio==0.23.0                 # Testing async functions
